requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
openai==1.12.0
pandas==2.2.0
python-docx==1.1.0
pypdf==4.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
pytest==8.0.1
httpx==0.26.0
